♻️ organize files and data structure
