+++ lib/cli/hashingMaster.js
