+++ lib/data/template.json
@@ -10,6 +10,7 @@
     "rooms": [
       {
         "location": "string",
+        "title": "string",
         "start": "2020-10-01T10:00:00",
         "end": "2020-10-01T10:00:00",
         "default": true,
