--- lib/dataConverter/index.js
@@ -1,5 +1,0 @@
-const { buildDailyOverview } = require("./dailyOverview");
-const { createFunctionSheet } = require("./functionSheet");
-
-exports.buildDailyOverview = buildDailyOverview;
-exports.createFunctionSheet = createFunctionSheet;
