--- lib/timeHandler/index.js
@@ -1,5 +1,0 @@
-const { createTime } = require("./createTimeforSheet");
-const { createTimeScale } = require("./timestampConverter");
-
-exports.createTime = createTime;
-exports.createTimeScale = createTimeScale;
