+++ lib/utils/convertTimestamp.js
