+++ lib/utils/convertToDailyOverview.js
@@ -1,4 +1,4 @@
-const { createTime } = require("../timeHandler/");
+const { createTime } = require(".");
 
 const buildDailyOverview = (documents, { now }) => {
   // delay enabled to filter events up to 6 hour
