+++ lib/utils/convertToFunctionSheet.js
@@ -1,4 +1,4 @@
-const { createTime } = require("../timeHandler/");
+const { createTime } = require(".");
 
 function createFunctionSheet(document) {
   const sheet = [
