+++ lib/utils/createPeriodOfTime.js
@@ -1,6 +1,6 @@
 const isUnixTime = (date) => Math.floor(date / 1000);
 
-function createTimeScale(timestamp) {
+function createPeriodOfTime(timestamp) {
   const givenDate = new Date(timestamp * 1000);
   const start = isUnixTime(
     new Date(
@@ -28,4 +28,4 @@ function createTimeScale(timestamp) {
 }
 
 exports.isUnixTime = isUnixTime;
-exports.createTimeScale = createTimeScale;
+exports.createPeriodOfTime = createPeriodOfTime;
